+++ src/Atata.TestSuite/Class.cs
@@ -1,0 +1,9 @@
+﻿namespace $rootnamespace$;
+
+public sealed class $safeitemrootname$ : AtataTestSuite
+{
+	[Test]
+	public void Test()
+	{
+	}
+}