+++ src/Atata.TestSuite/Properties/AssemblyInfo.cs
@@ -1,6 +1,6 @@
 ﻿using System.Reflection;
 using System.Runtime.InteropServices;
 
-[assembly: AssemblyTitle("Atata.NUnitTestFixture")]
+[assembly: AssemblyTitle("Atata.TestSuite")]
 [assembly: AssemblyDescription("")]
 [assembly: Guid("6ebeb766-810a-4807-b23f-9d33f78f2e40")]